Add support for metadata in definitions
All nodes in a definition apart from the last one are considered
"metadata" and are read in at resolve time.

 - A normal string defines the doc string for a function. There can only
   be one doc string for a definition.
 - A key defines a modifier. Currently the only modifier is "hidden":
   which prevents a definition from being exported.

The copy-meta macro from base.lisp will extract metadata from the body
of a function definition (such as defun or defmacro) and push it into
the main definition, converting

(defun my-fun ()
  "My function does something"
  23)

into

(define my-fun
  "My function does something"
  (lambda () 23))

It handles both table and string values, and ensures that there is at
least one expression remaining in the function body.