Add support for top level unquoting
Top level quoting allows executing arbitrary code at compile time
without the requirement of a macro. It's implementation is much the same
- it creates a compilation state, resolves the node and executes it. The
resulting value is then substituted back in.

There is a small difference in that expressions are wrapped in lambdas,
meaning they can be invoked when resolving rather than compiling. This
allows us to deal with errors in a nicer way, even if it is a bit of a
hack.