Use const-struct instead of struct in parser
struct attempts to coerce from keys to strings.  Whilst this is a good
idea in principle, it does have a detrimental effect on performance. We
add two new definitions here and use them inside the parser:

 - fast-struct will not coerce keys in entries. This makes it about
   twice as fast as struct. However, this does mean it doesn't work with
   quoted values.
 - const-struct generates the table at compile time, making it four
   times as fast as struct. However, this does require you to know all
   keys at compile time.
   Due to the nature of it's definition, it may also create a
   closure. We may be able to eliminate this in code gen, but it is
   quicker than allocating a list and looping over that.

This brings the parser time down from 1.8 seconds to 1.3 seconds on my
machine. Mileage may vary.