Sensibly compile instantiations of structs
This detects expressions of the form:

    ((lambda (x) (set-idx! x A B)... x) (empty-struct))

And compiles them to the appropriate table constructor. Whilst this
gives small performance gains, it also relies on empty-struct and
get-idx! being their original definitions and will not work if using an
alias of the symbol.

Hopefully we will be able to move this out into some form of compiler
extension once we get a system to enable that.