Inline directly called lambdas in more cases
In the case where arguments are executed in order, and no impure code
executes before they are evaluated, then this function is considered
suitable for merging.

Doing this naievly results in issues with variable return functions, so
we wrap values which might return multiple values in a directly called
lambda which limits it to one value. The codegen can then detect these
cases and emit more efficient code as appropriate.