Suggest similar variables when one cannot be resolved
We use an adapted Levenshtein distance to find "similar" names in
scope. We have to make some adaptions in order to handle short strings,
such as function operators, as otherwise pretty much every operator is
considered similar.

There are probably some improvements which can be made here, such as
handling imported modules in a more intelligent way (or even searching
loaded modules for such a symbol).