Add rudimentary support for thousands separators
I chose ' to be the thousands separator: This does not conflict with
quote in any level but conceptual due to how the language is parsed.

The current implementation allows things like 1''0 and 1'0'0'0. Do we
want this? Should we allow only well-formed numbers (1'000'000)? Should
we only allow this in decimal numbers, or do we allow it in all bases?
(The current implementation allows silly things like #rM'DCL'XVI)