Rudimentary support for using Nix as an Urn package manager
Yes, I can do this.

The compiler derivation isn't changed a lot: We only take a list of
extra dependencies (which must have a .libraryPath attribute) which
are added to the compiler search path.

The interesting bits (i.e. handling dependencies, etc) are all done by
urn.nix[1], which needs to be placed in the same directory as a clone of
the compiler (as compiler/) due to silly reasons.

Then, you can write Urn packages like so:

  import ../urn.nix {
    name = "foo";
    version = "0";
    src = ./.;

    executable = false;
  }

This is a package that doesn't depend on anything except for the
compiler, and builds a library (executable = false -> library = true).
A package which depends on this one might be specified like

  import ../urn.nix {
    name = "bar";
    src = ./.;

    depends = [ ../foo/default.nix ];
    executable = true;
  }

The Nix urn infrastructure will arrange for foo to be built (which boils
down to copying the Lisp files across), and then for a specialised
compiler binary to be generated which has foo in its lookup path, which
is then used to build bar.

Note that we don't support remote packages yet (you can get around this
by making your derivation depend on fetching the derivations with
nativeDepends).

1: https://hydraz.club/txt/urn.nix /
   https://hydraz.club/txt/urn.nix.html