Add some infrastructure for generic methods
Probably ought to elaborate on this one:
The multimethods as implemented here implement lookup as a tree with no
backtracking, so that `(defmethod (foo x y) ...)` will introduce `(.>
foo :lookup x y)` as a method entry. This means that, for the simplest
case of single dispath on a single argument, lookup is constant-time.

There is currently no way to specify on which arguments dispatch should
start (so, for instance, all specialisations of `(defgeneric map (f
xs))` would need to be `(defmethod (map function ...) ...)` instead of
merely `(defmethod (map ...) ...)`), or to only do dispatch on a single
argument.