Massive improvements to c[ad]+r generation
This actually generates the "correct" code now, rather than whatever
mess it generated before. It also will delegate subsequent accesses to
other c[ad]+r methods to ensure better runtime type safety.

I'll be the first to admit that the code is rather ugly. I'm open to
suggestions on how to improve it.