Add a unified transformer for optimisation passes
Instead of walking the same tree several times, we walk it once,
calling various visitor nodes where appropriate. This reduces the
overhead of visiting and means optimisations can be applied
subsequently.

Furthermore, the transformer will automatically update variable
definitions where required, meaning we have a higher chance of nodes
being reduced to constants and being propagatable.

In order to generate a list of valid transformations, we bake a list of
optimisation passes before hand. This is also done for warning passes.