Defensively copy captured loop variables
Consider some expression like so:

(loop [(x 1)] []
  (foo (lambda (x)))
  (recur x))

As the loop variable `x` is captured, our normal translation is not valid
as the upvalue will also be mutated. Therefore we need to defensively
copy values which are captured.

This is done as part of node categorisation, in a small analysis pass on
all recursive nodes. Inside a loop body we swap out captured variables
with a temporary variable, reverting to the original node when the loop
body ends.

Note there are a couple of edge cases which are not handled (loop
variables being captured in a condition) but that can be fixed at a later
date when I'm feeling less sleepy.